requirements.txt
google-generativeai
chromadb
loguru
python-dotenv
tqdm
beautifulsoup4
lxml
requests
fastapi
uvicorn
